refactored out bower_components
